pyproject.toml
[tool.poetry]
name = "auto-report-agent"
version = "0.1.0"
description = ""
authors = ["Akihiro Inui"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mongo = "^0.2.0"
python-decouple = "^3.8"
openai = "^1.93.0"
azure-identity = "^1.23.0"
openai-agents = "^0.1.0"
pytz = "^2025.2"

[tool.poetry.group.mcp.dependencies]
fastapi = "^0.115.14"
uvicorn = "^0.35.0"
mcp = {extras = ["cli"], version = "^1.10.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
